requirements.txt
spellchecker
bs4
huggingface_hub
language_tool_python
symspellpy
docling
utils
transformers
#flash_attn --no-build-isolation
faiss-cpu
lancedb
langchain_community
langchain_docling
langchain_huggingface
langchain_milvus